splinting is now up to scicos module

+++ scilab/modules/scicos_blocks/src/c/extract_bit_16_RB0.c
@@ -37,7 +37,7 @@ SCICOS_BLOCKS_IMPEXP void extract_bit_16_RB0(scicos_block *block, int flag)
     numb = *(ipar + 1) - *ipar + 1;
     for (i = 0; i < numb; i++)
     {
-        n = (short)pow(2, *ipar + i);
+        n = (short)pow(2, (double)(*ipar + i));
         ref = ref + n;
     }
     *y = (*u) & (ref);
