+++ scilab/modules/scicos_blocks/src/c/extract_bit_8_LH.c
@@ -35,7 +35,7 @@ SCICOS_BLOCKS_IMPEXP void extract_bit_8_LH(scicos_block *block, int flag)
     ref = 0;
     for (i = 0; i < maxim / 2; i++)
     {
-        n = (char)pow(2, i);
+        n = (char)pow(2, (double)i);
         ref = ref + n;
     }
     *y = (*u) & (ref);
